requirements.txt
langchain-core
langchain
langchain-community
langchain-text-splitters
langchain-groq
langgraph
chromadb
streamlit
pypdf
packaging